+++ R/center.R
@@ -2,10 +2,7 @@
 #'
 #' Performs a grand-mean centering of data.
 #'
-#' @param x A data frame, a (numeric or character) vector or a factor.
-#' @param select Character vector of column names. If `NULL` (the default),
-#'   all variables will be selected.
-#' @param exclude Character vector of column names to be excluded from selection.
+#' @param x A (grouped) data frame, a (numeric or character) vector or a factor.
 #' @param force Logical, if `TRUE`, forces centering of factors as
 #'   well. Factors are converted to numerical values, with the lowest level
 #'   being the value `1` (unless the factor has numeric levels, which are
@@ -32,6 +29,7 @@
 #'   order, unless a named vector is given. In this case, names are matched
 #'   against the names of the selected variables.
 #' @param ... Currently not used.
+#' @inheritParams find_columns
 #' @inheritParams standardize
 #'
 #' @note
@@ -86,9 +84,9 @@ centre <- center
 center.numeric <- function(x,
                            robust = FALSE,
                            weights = NULL,
-                           verbose = TRUE,
                            reference = NULL,
                            center = NULL,
+                           verbose = TRUE,
                            ...) {
   args <- .process_std_center(x, weights, robust, verbose, reference, center, scale = NULL)
 
@@ -105,16 +103,17 @@ center.numeric <- function(x,
   attr(centered_x, "center") <- args$center
   attr(centered_x, "scale") <- args$scale
   attr(centered_x, "robust") <- robust
-  centered_x
+  # labels
+  .set_back_labels(centered_x, x, include_values = FALSE)
 }
 
 
 #' @export
 center.factor <- function(x,
-                          weights = NULL,
                           robust = FALSE,
-                          verbose = TRUE,
+                          weights = NULL,
                           force = FALSE,
+                          verbose = TRUE,
                           ...) {
   if (!force) {
     return(x)
@@ -135,15 +134,19 @@ center.character <- center.factor
 center.data.frame <- function(x,
                               robust = FALSE,
                               weights = NULL,
-                              verbose = TRUE,
                               reference = NULL,
+                              center = NULL,
+                              force = FALSE,
+                              remove_na = c("none", "selected", "all"),
+                              append = FALSE,
                               select = NULL,
                               exclude = NULL,
-                              remove_na = c("none", "selected", "all"),
-                              force = FALSE,
-                              append = FALSE,
-                              center = NULL,
+                              ignore_case = FALSE,
+                              verbose = TRUE,
                               ...) {
+  # evaluate select/exclude, may be select-helpers
+  select <- .select_nse(select, x, exclude, ignore_case, verbose = verbose)
+
   # process arguments
   args <- .process_std_args(x, select, exclude, weights, append,
     append_suffix = "_c", force, remove_na, reference,
@@ -177,15 +180,19 @@ center.data.frame <- function(x,
 center.grouped_df <- function(x,
                               robust = FALSE,
                               weights = NULL,
-                              verbose = TRUE,
                               reference = NULL,
+                              center = NULL,
+                              force = FALSE,
+                              remove_na = c("none", "selected", "all"),
+                              append = FALSE,
                               select = NULL,
                               exclude = NULL,
-                              remove_na = c("none", "selected", "all"),
-                              force = FALSE,
-                              append = FALSE,
-                              center = NULL,
+                              ignore_case = FALSE,
+                              verbose = TRUE,
                               ...) {
+  # evaluate select/exclude, may be select-helpers
+  select <- .select_nse(select, x, exclude, ignore_case, verbose = verbose)
+
   args <- .process_grouped_df(x, select, exclude, append,
     append_suffix = "_c",
     reference, weights, force