+++ R/data_reverse.R
@@ -2,12 +2,11 @@
 #'
 #' Reverse-score variables (change the keying/scoring direction).
 #'
-#' @inheritParams standardize.data.frame
-#'
-#' @param x A numeric or factor variable.
 #' @param range Initial (old) range of values. If `NULL`, will take the range of
 #'   the input vector (`range(x)`).
 #' @param ... Arguments passed to or from other methods.
+#' @inheritParams data_cut
+#' @inheritParams find_columns
 #'
 #' @examples
 #' data_reverse(c(1, 2, 3, 4, 5))
@@ -28,8 +27,7 @@
 #'
 #' @family transform utilities
 #'
-#' @seealso [data_rescale()] to change the score range for variables (potentially while reversing),
-#'   [normalize()] [standardize()] [ranktransform()]
+#' @inherit data_rename seealso
 #'
 #' @export
 data_reverse <- function(x, ...) {
@@ -39,10 +37,7 @@ data_reverse <- function(x, ...) {
 
 #' @rdname data_reverse
 #' @export
-reverse_scale <- function(x, ...) {
-  # Alias for data_reverse()
-  data_reverse(x, ...)
-}
+reverse_scale <- data_reverse
 
 
 
@@ -62,7 +57,7 @@ data_reverse.numeric <- function(x,
   # Warning if only one value
   if (length(unique(x)) == 1 && is.null(range)) {
     if (verbose) {
-      warning(paste0("A `range` must be provided for data with only one unique value."))
+      warning("A `range` must be provided for data with only one unique value.", call. = FALSE)
     }
     return(x)
   }
@@ -77,6 +72,9 @@ data_reverse.numeric <- function(x,
   new_max <- min
 
   out <- as.vector((new_max - new_min) / (max - min) * (x - min) + new_min)
+
+  # labelled data?
+  out <- .set_back_labels(out, x)
   out
 }
 
@@ -94,10 +92,13 @@ data_reverse.factor <- function(x, range = NULL, verbose = TRUE, ...) {
   # Warning if only one value
   if (length(unique(x)) == 1 && is.null(range)) {
     if (verbose) {
-      warning("A `range` must be provided for data with only one unique value.")
+      warning("A `range` must be provided for data with only one unique value.", call. = FALSE)
     }
     return(x)
   }
+
+  # save for later use
+  original_x <- x
 
   if (!is.null(range)) {
     old_levels <- range
@@ -109,31 +110,30 @@ data_reverse.factor <- function(x, range = NULL, verbose = TRUE, ...) {
   int_x <- as.integer(x)
   rev_x <- data_reverse(int_x, range = c(1, length(old_levels)))
   x <- factor(rev_x, levels = seq_len(length(old_levels)), labels = old_levels)
+
+  # labelled data?
+  x <- .set_back_labels(x, original_x)
+
   x
 }
 
 
 
 
-#' @rdname data_reverse
 #' @export
 data_reverse.grouped_df <- function(x,
                                     range = NULL,
                                     select = NULL,
                                     exclude = NULL,
+                                    ignore_case = FALSE,
                                     ...) {
   info <- attributes(x)
 
   # dplyr >= 0.8.0 returns attribute "indices"
   grps <- attr(x, "groups", exact = TRUE)
 
-  # check for formula notation, convert to character vector
-  if (inherits(select, "formula")) {
-    select <- all.vars(select)
-  }
-  if (inherits(exclude, "formula")) {
-    exclude <- all.vars(exclude)
-  }
+  # evaluate arguments
+  select <- .select_nse(select, x, exclude, ignore_case)
 
   # dplyr < 0.8.0?
   if (is.null(grps)) {
@@ -165,23 +165,10 @@ data_reverse.data.frame <- function(x,
                                     range = NULL,
                                     select = NULL,
                                     exclude = NULL,
+                                    ignore_case = FALSE,
                                     ...) {
-
-  # check for formula notation, convert to character vector
-  if (inherits(select, "formula")) {
-    select <- all.vars(select)
-  }
-  if (inherits(exclude, "formula")) {
-    exclude <- all.vars(exclude)
-  }
-
-  if (is.null(select)) {
-    select <- names(x)
-  }
-
-  if (!is.null(exclude)) {
-    select <- setdiff(select, exclude)
-  }
+  # evaluate arguments
+  select <- .select_nse(select, x, exclude, ignore_case)
 
   # Transform the range so that it is a list now
   if (!is.null(range)) {
@@ -194,4 +181,18 @@ data_reverse.data.frame <- function(x,
     data_reverse(x[[n]], range = range[[n]])
   })
   x
+}
+
+
+
+# helper -----------------------------
+
+.set_back_labels <- function(new, old, include_values = TRUE) {
+  # labelled data?
+  attr(new, "label") <- attr(old, "label", exact = TRUE)
+  labels <- attr(old, "labels", exact = TRUE)
+  if (isTRUE(include_values) && !is.null(labels)) {
+    attr(new, "labels") <- stats::setNames(rev(labels), names(labels))
+  }
+  new
 }